--- a/pkg/streamingpromql/testdata/ours/functions.test
+++ b/pkg/streamingpromql/testdata/ours/functions.test
@@ -822,3 +822,20 @@ eval range from 0 to 7m step 1m year(some_metric{case="series1"} + another_metri
   {case="series1"} 1970 2022 2075 2128 2181 2234 2287 2340
 
 clear
+
+load 30s
+  some_metric{case="floats only"} 1 2 _ 3 4
+  some_metric{case="histograms only"} {{count:1}} {{count:2}} _ {{count:3}} {{count:4}}
+  some_metric{case="floats and histograms"} 1 2 _ {{count:3}} {{count:4}}
+
+# If timestamp() is used over a selector, it returns the timestamps of the underlying samples.
+eval range from 0 to 2m step 1m timestamp(some_metric)
+  {case="floats only"} 0 30 120
+  {case="histograms only"} 0 30 120
+  {case="floats and histograms"} 0 30 120
+
+# If timestamp() is used over something that is not a selector, it returns the timestamps of the points returned.
+eval range from 0 to 2m step 1m timestamp(some_metric * 2)
+  {case="floats only"} 0 60 120
+  {case="histograms only"} 0 60 120
+  {case="floats and histograms"} 0 60 120