--- a/pkg/streamingpromql/testdata/upstream/at_modifier.test
+++ b/pkg/streamingpromql/testdata/upstream/at_modifier.test
@@ -171,19 +171,16 @@ eval instant at 10s minute(metric @ 1500)
   {job="2"} 5
 
 # timestamp() takes the time of the sample and not the evaluation time.
-# Unsupported by streaming engine.
-# eval instant at 10m timestamp(metric{job="1"} @ 10)
-#   {job="1"} 10
+eval instant at 10m timestamp(metric{job="1"} @ 10)
+  {job="1"} 10
 
 # The result of inner timestamp() will have the timestamp as the
 # eval time, hence entire expression is not step invariant and depends on eval time.
-# Unsupported by streaming engine.
-# eval instant at 10m timestamp(timestamp(metric{job="1"} @ 10))
-#   {job="1"} 600
+eval instant at 10m timestamp(timestamp(metric{job="1"} @ 10))
+  {job="1"} 600
 
-# Unsupported by streaming engine.
-# eval instant at 15m timestamp(timestamp(metric{job="1"} @ 10))
-#   {job="1"} 900
+eval instant at 15m timestamp(timestamp(metric{job="1"} @ 10))
+  {job="1"} 900
 
 # Time functions inside a subquery.
 
@@ -213,16 +210,14 @@ eval instant at 0s sum_over_time(vector(time())[100s:1s] @ 3000 offset 600s)
   {} 235050
 
 # timestamp() takes the time of the sample and not the evaluation time.
-# Unsupported by streaming engine.
-# eval instant at 0s sum_over_time(timestamp(metric{job="1"} @ 10)[100s:10s] @ 3000)
-#   {job="1"} 100
+eval instant at 0s sum_over_time(timestamp(metric{job="1"} @ 10)[100s:10s] @ 3000)
+  {job="1"} 100
 
 # The result of inner timestamp() will have the timestamp as the
 # eval time, hence entire expression is not step invariant and depends on eval time.
 # Here eval time is determined by the subquery.
-# Unsupported by streaming engine.
-# eval instant at 0s sum_over_time(timestamp(timestamp(metric{job="1"} @ 999))[10s:1s] @ 10)
-#   {job="1"} 55
+eval instant at 0s sum_over_time(timestamp(timestamp(metric{job="1"} @ 999))[10s:1s] @ 10)
+  {job="1"} 55
 
 
 clear