--- a/pkg/streamingpromql/testdata/upstream/functions.test
+++ b/pkg/streamingpromql/testdata/upstream/functions.test
@@ -437,25 +437,20 @@ clear
 load 10s
   metric 1 1
 
-# Unsupported by streaming engine.
-# eval instant at 0s timestamp(metric)
-#   {} 0
+eval instant at 0s timestamp(metric)
+  {} 0
 
-# Unsupported by streaming engine.
-# eval instant at 5s timestamp(metric)
-#   {} 0
+eval instant at 5s timestamp(metric)
+  {} 0
 
-# Unsupported by streaming engine.
-# eval instant at 5s timestamp(((metric)))
-#   {} 0
+eval instant at 5s timestamp(((metric)))
+  {} 0
 
-# Unsupported by streaming engine.
-# eval instant at 10s timestamp(metric)
-#   {} 10
+eval instant at 10s timestamp(metric)
+  {} 10
 
-# Unsupported by streaming engine.
-# eval instant at 10s timestamp(((metric)))
-#   {} 10
+eval instant at 10s timestamp(((metric)))
+  {} 10
 
 # Tests for label_join.
 load 5m
@@ -1604,9 +1599,8 @@ load 1m
   metric 0+1x1000
 
 # We expect the value to be 0 for t=0s to t=59s (inclusive), then 60 for t=60s and t=61s.
-# Unsupported by streaming engine.
-# eval range from 0 to 61s step 1s timestamp(metric)
-#   {} 0x59 60 60
+eval range from 0 to 61s step 1s timestamp(metric)
+  {} 0x59 60 60
 
 clear